Extract votes package
