+++ src/main/java/com/code4ro/legalconsultation/vote/model/persistence/Vote.java
@@ -1,6 +1,8 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.vote.model.persistence;
 
 import com.code4ro.legalconsultation.comment.model.persistence.Comment;
+import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
