+++ src/test/java/com/code4ro/legalconsultation/vote/service/VoteServiceImplGetAllTest.java
@@ -1,10 +1,13 @@
-package com.code4ro.legalconsultation.service.impl;
+package com.code4ro.legalconsultation.vote.service;
 
 import com.code4ro.legalconsultation.comment.model.persistence.Comment;
 import com.code4ro.legalconsultation.config.security.CurrentUserService;
-import com.code4ro.legalconsultation.model.dto.VoteDto;
+import com.code4ro.legalconsultation.vote.model.dto.VoteDto;
 import com.code4ro.legalconsultation.model.persistence.*;
-import com.code4ro.legalconsultation.repository.VoteRepository;
+import com.code4ro.legalconsultation.vote.repository.VoteRepository;
+import com.code4ro.legalconsultation.vote.model.persistence.Vote;
+import com.code4ro.legalconsultation.vote.model.persistence.VoteType;
+import com.code4ro.legalconsultation.vote.service.impl.VoteServiceImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -19,8 +22,8 @@ import java.util.UUID;
 
 import static com.code4ro.legalconsultation.factory.RandomObjectFiller.createAndFill;
 import static com.code4ro.legalconsultation.user.model.persistence.UserRole.ADMIN;
-import static com.code4ro.legalconsultation.model.persistence.VoteType.DOWN;
-import static com.code4ro.legalconsultation.model.persistence.VoteType.UP;
+import static com.code4ro.legalconsultation.vote.model.persistence.VoteType.DOWN;
+import static com.code4ro.legalconsultation.vote.model.persistence.VoteType.UP;
 import static java.util.UUID.randomUUID;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
